+++ example/blog/views.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def search_page(request):
+    return render(request, "search.html")